+++ src/main/java/com/liquibaseproject/entity/Users.java
@@ -2,11 +2,24 @@ package com.liquibaseproject.entity;
 
 import jakarta.persistence.*;
 
+import java.sql.Timestamp;
 import java.util.Date;
 import java.util.UUID;
 
 @Entity
 @Table(name = "users")
+@NamedStoredProcedureQuery(
+        name = "insertUser",
+        procedureName = "insert_user",
+        parameters = {
+                @StoredProcedureParameter(mode = ParameterMode.IN, name = "p_username", type = String.class),
+                @StoredProcedureParameter(mode = ParameterMode.IN, name = "p_password", type = String.class),
+                @StoredProcedureParameter(mode = ParameterMode.IN, name = "p_email", type = String.class),
+                @StoredProcedureParameter(mode = ParameterMode.OUT, name = "o_username", type = String.class),
+                @StoredProcedureParameter(mode = ParameterMode.OUT, name = "o_created_at", type = Timestamp.class),
+                @StoredProcedureParameter(mode = ParameterMode.OUT, name = "o_updated_at", type = Timestamp.class)
+        }
+)
 public class Users {
 
     @Id